+++ src/main/java/com/back/code_brew/domain/cart/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package com.back.code_brew.domain.cart.repository;
 import com.back.code_brew.domain.cart.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 }
